+++ stations.js
@@ -1,13 +1,13 @@
 export const stations = [
-    { name: "ETN-FM Trance", url: "https://stream.pcradio.ru/etnfm_trance-hi" },
-    { name: "Psyndora Psytrance", url: "https://cast.magicstreams.gr/sc/psyndora/stream" },
-    { name: "DMT-FM", url: "https://dc1.serverse.com/proxy/ywycfrxn/stream" },
-    { name: "BOM Psytrance Radio", url: "https://strm112.1.fm/psytrance_mobile_mp3?aw_0_req.gdpr=true" },
-    { name: "Psychedelic FM", url: "https://s2.radio.co/s2696f08b5/listen" },
-    { name: "Melodic House Radio", url: "https://stream1-technolovers.radiohost.de/melodic-house-techno" },
-    { name: "psyradio * fm - progressive", url: "http://65.109.32.21:8010/stream" },
-    { name: "Party~Vibe | Psychedelic Trance Radio", url: "http://94.130.242.5:8010/stream3" },
-    { name: "DI.FM - Vocal Trance", url: "https://listen.di.fm/vocaltrance" },
-    { name: "Radiozora - Psytrance", url: "https://2026.radiozora.fm/stream/" },
-    { name: "A State of Trance (ASOT)", url: "https://streaming.radio.co/s0cf4c8c63/listen" }
+    { name: "ETN-FM Trance", url: "https://stream.pcradio.ru/etnfm_trance-hi", genre: "Trance", country: "Canada" },
+    { name: "Psyndora Psytrance", url: "https://cast.magicstreams.gr/sc/psyndora/stream", genre: "Psytrance", country: "Greece" },
+    { name: "DMT-FM", url: "https://dc1.serverse.com/proxy/ywycfrxn/stream", genre: "Psytrance", country: "UK" },
+    { name: "BOM Psytrance Radio", url: "https://strm112.1.fm/psytrance_mobile_mp3?aw_0_req.gdpr=true", genre: "Psytrance", country: "Switzerland" },
+    { name: "Psychedelic FM", url: "https://s2.radio.co/s2696f08b5/listen", genre: "Psytrance", country: "Netherlands" },
+    { name: "Melodic House Radio", url: "https://stream1-technolovers.radiohost.de/melodic-house-techno", genre: "Melodic House", country: "Germany" },
+    { name: "psyradio * fm - progressive", url: "http://65.109.32.21:8010/stream", genre: "Progressive Psy", country: "Germany" },
+    { name: "Party~Vibe | Psychedelic Trance Radio", url: "http://94.130.242.5:8010/stream3", genre: "Psytrance", country: "UK" },
+    { name: "DI.FM - Vocal Trance", url: "https://listen.di.fm/vocaltrance", genre: "Vocal Trance", country: "USA" },
+    { name: "Radiozora - Psytrance", url: "https://2026.radiozora.fm/stream/", genre: "Psytrance", country: "Hungary" },
+    { name: "A State of Trance (ASOT)", url: "https://streaming.radio.co/s0cf4c8c63/listen", genre: "Trance", country: "Netherlands" }
 ];